moving UI specific code into its own folder
